backend-golocal/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
cloudinary==1.35.0
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
load-dotenv==0.1.0
mysqlclient==2.2.0
Pillow==10.0.1
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.16
